feat(trusty-mpm): port language engineers + QA variants into bundled assets (#769)
Part of #769 — Phase 1 increment 2: language engineers + QA variants

Adds 12 concrete agents to `bundle::ALL` (count: 20 → 32):

Language engineers (extends: base-engineer, model: sonnet):
  python-engineer, typescript-engineer, golang-engineer, rust-engineer,
  java-engineer, php-engineer, ruby-engineer, react-engineer,
  nextjs-engineer, svelte-engineer

QA variants (extends: base-qa, model: sonnet):
  web-qa (6-phase progressive + UAT mode), api-qa (REST/GraphQL/contract)

Each agent uses the project's flat 5-field frontmatter format (name, role,
description, model, extends). bundle_tests.rs is split out of bundle.rs to
keep both files under the 500-line hard cap; the module is wired via
`#[cfg(test)] #[path = "bundle_tests.rs"] mod tests;`.

Quality bar: cargo build, test (830 tests pass), clippy -D warnings,
fmt --check, check_line_cap.sh — all clean.
