+++ crates/trusty-mpm/src/core/bundle.rs
@@ -77,6 +77,42 @@ pub const TICKETING_AGENT: &str = include_str!("../assets/agents/ticketing.md");
 /// Concrete code-analyzer agent (`extends: base-research`).
 pub const CODE_ANALYZER_AGENT: &str = include_str!("../assets/agents/code-analyzer.md");
 
+/// Concrete python-engineer agent (`extends: base-engineer`).
+pub const PYTHON_ENGINEER_AGENT: &str = include_str!("../assets/agents/python-engineer.md");
+
+/// Concrete typescript-engineer agent (`extends: base-engineer`).
+pub const TYPESCRIPT_ENGINEER_AGENT: &str = include_str!("../assets/agents/typescript-engineer.md");
+
+/// Concrete golang-engineer agent (`extends: base-engineer`).
+pub const GOLANG_ENGINEER_AGENT: &str = include_str!("../assets/agents/golang-engineer.md");
+
+/// Concrete rust-engineer agent (`extends: base-engineer`).
+pub const RUST_ENGINEER_AGENT: &str = include_str!("../assets/agents/rust-engineer.md");
+
+/// Concrete java-engineer agent (`extends: base-engineer`).
+pub const JAVA_ENGINEER_AGENT: &str = include_str!("../assets/agents/java-engineer.md");
+
+/// Concrete php-engineer agent (`extends: base-engineer`).
+pub const PHP_ENGINEER_AGENT: &str = include_str!("../assets/agents/php-engineer.md");
+
+/// Concrete ruby-engineer agent (`extends: base-engineer`).
+pub const RUBY_ENGINEER_AGENT: &str = include_str!("../assets/agents/ruby-engineer.md");
+
+/// Concrete react-engineer agent (`extends: base-engineer`).
+pub const REACT_ENGINEER_AGENT: &str = include_str!("../assets/agents/react-engineer.md");
+
+/// Concrete nextjs-engineer agent (`extends: base-engineer`).
+pub const NEXTJS_ENGINEER_AGENT: &str = include_str!("../assets/agents/nextjs-engineer.md");
+
+/// Concrete svelte-engineer agent (`extends: base-engineer`).
+pub const SVELTE_ENGINEER_AGENT: &str = include_str!("../assets/agents/svelte-engineer.md");
+
+/// Concrete web-qa agent (`extends: base-qa`).
+pub const WEB_QA_AGENT: &str = include_str!("../assets/agents/web-qa.md");
+
+/// Concrete api-qa agent (`extends: base-qa`).
+pub const API_QA_AGENT: &str = include_str!("../assets/agents/api-qa.md");
+
 /// Placeholder skill definition installed to `skills/example-skill.md`.
 pub const EXAMPLE_SKILL: &str = include_str!("../assets/skills/example-skill.md");
 
@@ -225,6 +261,66 @@ pub const ALL: &[BundledArtifact] = &[
         install: InstallPolicy::Overwrite,
     },
     BundledArtifact {
+        rel_path: "agents/python-engineer.md",
+        contents: PYTHON_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/typescript-engineer.md",
+        contents: TYPESCRIPT_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/golang-engineer.md",
+        contents: GOLANG_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/rust-engineer.md",
+        contents: RUST_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/java-engineer.md",
+        contents: JAVA_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/php-engineer.md",
+        contents: PHP_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/ruby-engineer.md",
+        contents: RUBY_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/react-engineer.md",
+        contents: REACT_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/nextjs-engineer.md",
+        contents: NEXTJS_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/svelte-engineer.md",
+        contents: SVELTE_ENGINEER_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/web-qa.md",
+        contents: WEB_QA_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
+        rel_path: "agents/api-qa.md",
+        contents: API_QA_AGENT,
+        install: InstallPolicy::Overwrite,
+    },
+    BundledArtifact {
         rel_path: "skills/example-skill.md",
         contents: EXAMPLE_SKILL,
         install: InstallPolicy::Overwrite,
@@ -232,224 +328,5 @@ pub const ALL: &[BundledArtifact] = &[
 ];
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn constants_are_non_empty() {
-        // Every embedded artifact must carry real content — an empty
-        // `include_str!` target would mean a missing or truncated asset file.
-        assert!(!OPTIMIZER_TOML.trim().is_empty());
-        assert!(!OVERSEER_TOML.trim().is_empty());
-        assert!(!FRAMEWORK_INSTRUCTIONS.trim().is_empty());
-        assert!(!CLAUDE_STUB.trim().is_empty());
-        assert!(!BASE_AGENT.trim().is_empty());
-        assert!(!BASE_ENGINEER.trim().is_empty());
-        assert!(!BASE_RESEARCH.trim().is_empty());
-        assert!(!BASE_QA.trim().is_empty());
-        assert!(!BASE_OPS.trim().is_empty());
-        assert!(!ENGINEER_AGENT.trim().is_empty());
-        assert!(!QA_AGENT.trim().is_empty());
-        assert!(!RESEARCH_AGENT.trim().is_empty());
-        assert!(!OPS_AGENT.trim().is_empty());
-        assert!(!SECURITY_AGENT.trim().is_empty());
-        assert!(!DOCUMENTATION_AGENT.trim().is_empty());
-        assert!(!DATA_ENGINEER_AGENT.trim().is_empty());
-        assert!(!VERSION_CONTROL_AGENT.trim().is_empty());
-        assert!(!TICKETING_AGENT.trim().is_empty());
-        assert!(!CODE_ANALYZER_AGENT.trim().is_empty());
-        assert!(!EXAMPLE_SKILL.trim().is_empty());
-        assert!(!OUTPUT_STYLE.trim().is_empty());
-    }
-
-    #[test]
-    fn output_style_has_matching_frontmatter_name() {
-        // Claude Code matches the `outputStyle` settings key against the
-        // `name:` in the style file's frontmatter; a mismatch silently falls
-        // back to the operator's default style.
-        assert!(OUTPUT_STYLE.contains("name: trusty-mpm"));
-    }
-
-    #[test]
-    fn framework_instructions_and_stub_differ() {
-        // The framework artifact and the user stub are distinct files with
-        // distinct content; conflating them would lose either upgrades or
-        // user edits.
-        assert_ne!(FRAMEWORK_INSTRUCTIONS, CLAUDE_STUB);
-    }
-
-    #[test]
-    fn claude_stub_is_seed_once() {
-        // The user stub must never be overwritten on re-install.
-        let stub = ALL
-            .iter()
-            .find(|a| a.rel_path == "instructions/CLAUDE.md")
-            .expect("CLAUDE.md stub present in bundle");
-        assert_eq!(stub.install, InstallPolicy::SeedOnce);
-    }
-
-    #[test]
-    fn framework_instructions_overwrites() {
-        // The framework instructions must be refreshed on every install.
-        let instr = ALL
-            .iter()
-            .find(|a| a.rel_path == "instructions/INSTRUCTIONS.md")
-            .expect("INSTRUCTIONS.md present in bundle");
-        assert_eq!(instr.install, InstallPolicy::Overwrite);
-    }
-
-    #[test]
-    fn optimizer_toml_is_parseable() {
-        // The shipped policy must be valid TOML or the installer would deploy
-        // a file the daemon then fails to load.
-        let parsed: toml::Value = toml::from_str(OPTIMIZER_TOML).expect("valid TOML");
-        assert!(parsed.get("default").is_some());
-    }
-
-    #[test]
-    fn bundle_table_is_complete() {
-        // `ALL` must enumerate every artifact with unique, non-empty paths.
-        // Count: 4 hooks/instructions + 5 base agents + 10 concrete agents + 1 skill = 20
-        // (hooks/optimizer.toml, hooks/overseer.toml, instructions/INSTRUCTIONS.md,
-        //  instructions/CLAUDE.md, agents/BASE-AGENT.md, agents/BASE-ENGINEER.md,
-        //  agents/BASE-RESEARCH.md, agents/BASE-QA.md, agents/BASE-OPS.md,
-        //  agents/engineer.md, agents/qa.md, agents/research.md, agents/ops.md,
-        //  agents/security.md, agents/documentation.md, agents/data-engineer.md,
-        //  agents/version-control.md, agents/ticketing.md, agents/code-analyzer.md,
-        //  skills/example-skill.md)
-        assert_eq!(ALL.len(), 20);
-        let mut paths: Vec<&str> = ALL.iter().map(|a| a.rel_path).collect();
-        paths.sort_unstable();
-        paths.dedup();
-        assert_eq!(paths.len(), 20, "artifact paths must be unique");
-        for artifact in ALL {
-            assert!(!artifact.rel_path.is_empty());
-            assert!(!artifact.contents.trim().is_empty());
-        }
-    }
-
-    #[test]
-    fn overseer_toml_is_parseable() {
-        // The shipped overseer policy must be valid TOML and ship disabled —
-        // oversight is opt-in, so installing it must not silently enable it.
-        let parsed: toml::Value = toml::from_str(OVERSEER_TOML).expect("valid TOML");
-        assert_eq!(
-            parsed
-                .get("overseer")
-                .and_then(|o| o.get("enabled"))
-                .and_then(toml::Value::as_bool),
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn overseer_toml_is_in_bundle() {
-        // `ALL` must include the overseer policy so `trusty-mpm install`
-        // deploys it.
-        assert!(
-            ALL.iter().any(|a| a.rel_path == "hooks/overseer.toml"),
-            "overseer.toml must be a bundled artifact"
-        );
-    }
-
-    #[test]
-    fn new_concrete_agents_are_in_bundle() {
-        // Every newly-ported concrete agent must be present in ALL so
-        // `trusty-mpm install` deploys them offline.
-        let agent_paths: Vec<&str> = ALL
-            .iter()
-            .filter(|a| a.rel_path.starts_with("agents/"))
-            .map(|a| a.rel_path)
-            .collect();
-
-        for expected in &[
-            "agents/qa.md",
-            "agents/research.md",
-            "agents/ops.md",
-            "agents/security.md",
-            "agents/documentation.md",
-            "agents/data-engineer.md",
-            "agents/version-control.md",
-            "agents/ticketing.md",
-            "agents/code-analyzer.md",
-        ] {
-            assert!(
-                agent_paths.contains(expected),
-                "missing bundled agent: {expected}"
-            );
-        }
-    }
-
-    #[test]
-    fn new_concrete_agents_have_extends_in_frontmatter() {
-        // Each new concrete agent must declare `extends:` so the inheritance
-        // chain resolves correctly at deploy time.
-        let agents = [
-            ("qa", QA_AGENT),
-            ("research", RESEARCH_AGENT),
-            ("ops", OPS_AGENT),
-            ("security", SECURITY_AGENT),
-            ("documentation", DOCUMENTATION_AGENT),
-            ("data-engineer", DATA_ENGINEER_AGENT),
-            ("version-control", VERSION_CONTROL_AGENT),
-            ("ticketing", TICKETING_AGENT),
-            ("code-analyzer", CODE_ANALYZER_AGENT),
-        ];
-        for (name, content) in agents {
-            assert!(
-                content.contains("extends:"),
-                "agent {name} is missing `extends:` in frontmatter"
-            );
-            // Composed output must not contain `extends:` (builder strips it).
-            // We verify the raw source has it; compose round-trip is covered
-            // by agent_deployer integration tests.
-        }
-    }
-
-    #[test]
-    fn new_concrete_agents_deploy_via_real_asset_files() {
-        // Verify each new agent composes without error using the real bundled
-        // asset files on disk (not temp fixtures) — catches typos in `extends:`
-        // values and missing base templates.
-        use crate::core::agent_builder::compose_agent;
-        use std::path::Path;
-
-        let assets_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("src")
-            .join("assets")
-            .join("agents");
-
-        let agents = [
-            "qa",
-            "research",
-            "ops",
-            "security",
-            "documentation",
-            "data-engineer",
-            "version-control",
-            "ticketing",
-            "code-analyzer",
-        ];
-
-        for name in agents {
-            let composed = compose_agent(name, &assets_dir)
-                .unwrap_or_else(|e| panic!("compose_agent({name}) failed: {e}"));
-            // Composed output must have a frontmatter block.
-            assert!(
-                composed.starts_with("---\n"),
-                "composed {name} is missing frontmatter"
-            );
-            // `extends:` must not leak into the composed output.
-            assert!(
-                !composed.contains("extends:"),
-                "composed {name} has leaked `extends:` in output"
-            );
-            // Must have non-trivial body content.
-            assert!(
-                composed.len() > 200,
-                "composed {name} suspiciously short ({} bytes)",
-                composed.len()
-            );
-        }
-    }
-}
+#[path = "bundle_tests.rs"]
+mod tests;
